Refactor for readability - reorganized functions into partial class files so I don't lose my mind.
Feat:
- Add atlas point indicators for league atlas points (breach, abyss, incursion, ritual, delirium)
- Add customization for atlas point indicators
