+++ .editorconfig
@@ -1,0 +1,3 @@
+root = true
+[*.cs]
+dotnet_diagnostic.IDE0005.severity = warning
